std: Add `process.getProcessId`.
This function is used to return an integer uniquely representing the
current process. Unlike process handles (which may be the same value
depending upon the operating system), the `process.ProcessId` is
intended for use by external actors within the same operating system.
The classic example being a "pid file". Another might be a process
specific checkpoint of the most recent input produced by libfuzzer (the
original motivating use case of this commit).