[DRAFT] libfuzzer: Add best-effort input persistence.
Previously, the fuzz input was backed by a regular allocator. Now, it is
backed by a memory-mapped file located at a path which can be
reconstructed by the parent process using the test process's PID. At
present that is only done on a best-effort basis which should be
sufficient to recover from a crashed run.